services/ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
mlflow==2.10.0

# OpenTelemetry (tracing distribué) — opentelemetry-exporter-jaeger-thrift 1.22+ non publié pour Python 3.14, utiliser 1.21.x
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-jaeger-thrift>=1.20.0,<1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-semantic-conventions>=0.43b0
